ECS remake, Parent & Children hierarchy, UI components
Did a remake of the ECS to remove templates, which increases compile time considerably; it will allow for addition of more components without increasing compile time exponentially.

Entities can now have parents and children which will be useful for some component types.

I have also implemented buttons and an early version of Panels, as a start of the UI system.